Move async_test to helper for reuse
